Handled new import and replace of OBS Default audio
